+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KHBM60MZABXBWBVA1MKQ3KS1"
+        "01KHBX71K5YZ42QS2K16YYGZFA"
     )
     r"""The unique identifier of the data source"""