+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -148,7 +148,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KHBM60MT9HZXRRPWNV2D1T8K"
+        "01KHBX71K175RMQWB1PNCK84S3"
     )
     r"""The unique identifier of the data source"""