+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KHBM60MXQMTC68PPX6CT27AF"
+        "01KHBX71K4G77TQTDSSBZHJFKK"
     )
     r"""The unique identifier of the data source"""