+++ packages/orq-rc/RELEASES.md
@@ -5649,3 +5649,13 @@ Based on:
 - [python v4.3.0-rc.65] packages/orq-rc
 ### Releases
 - [PyPI v4.3.0-rc.65] https://pypi.org/project/orq-ai-sdk/4.3.0-rc.65 - packages/orq-rc
+
+## 2026-02-13 16:29:41
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.715.0 (2.822.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.3.0-rc.66] packages/orq-rc
+### Releases
+- [PyPI v4.3.0-rc.66] https://pypi.org/project/orq-ai-sdk/4.3.0-rc.66 - packages/orq-rc